Add and switch to new types for revocation
Add the UInt48, RevocationKey, CommitmentPubKey, CommitmentSeed and
RevocationSet types.

A RevocationKey is a secret key that can be used to revoke a
transaction. It's shared in the revoke_and_ack message and stored in the
user's wallet in a RevocationSet. A CommitmentPubKey is the public key
of a RevocationKey, also known as a "per commitment point". A
CommitmentSeed is the master revocation key for a side of a channel and
can be used to derive all the revocation keys for that side of that
channel.

A RevocationSet stores a set of revocation keys in a compact form
specified in BOLT #3 which allows earlier keys to be derived from later
keys.