Implement revocation (#11)
This commit adds the following:

* getFundsFromForceClosingTransaction

This function takes an on-chain transaction which spends the channel
funds and tries to extract spendable utxos out of it. If successful, it
returns a TransactionBuilder with txins added for each spendable output
of the closing transaction and with the necessary keys and
BuilderExtension added. To recover funds from a broadcast commitment
transaction you just need to call this function, add outputs to the
returned TransactionBuilder to send the money where you want it to go,
then broadcast the transaction.

* CommitmentToLocalBuilderExtension

This is an NBitcoin BuilderExtension that tells TransactionBuilder how
to recognise and sign lightning commitment transaction to_local outputs.
getFundsFromForceClosingTransaction adds this extension to the
TransactionBuilder it returns if it's needed to sign the transaction.

* SeqConsumer

A computation expression which makes it easy to write code that consumes
a sequence one element at a time.

* OptionCE

A computation expression for creation options. Similar to the result
computation expression.