Remove NormalData.Buried field
It's not at all clear what this field was intended for. If it's meant to
indicate the the funding tx is buried then we should never be in a
`Normal` state with `Buried = false` since being in a `Normal` state is
meant to indicate that funding has been locked.

At any rate, the field was never being read so I've removed it.