Merge WaitForFunding{Locked,Confirmed} states into Normal
Rather than having separate `ChannelState`s for waiting for funding
confirmed/locked we can merge these into the Normal state. The Normal
state's ShortChannelId and RemoteNextCommitInfo have been turned into
options where a value of `None` indicates that funding is not confirmed
or that the peer has not sent funding_locked respectively.

The advantages of this are:
  (a) It simplifies the code by removing two more channel states.
  (b) it allows us to handle channel shutdown before funding has been
      locked. This is required by the lightning spec but, prior to this
      change, attempting to apply a remote's `shutdown` message before
      the funding is confirmed and locked would have crashed DNL.
  (c) It's a step towards refactoring the channel type into two layers -
      one which is saved to the wallet and persisted across reconnects,
      and another which is discarded when the connection is dropped.