Remove ClosingTxProposed type
Remove ClosingTxProposed type and change the ClosingTxProposed field of
NegotiatingData to just be a list of proposed fees instead.

We don't need to store copies of the closing_signed messages and nor the
closing transactions that we've proposed. We don't currently use them
and we can always re-create them on demand anyway.