Pass feerate and target address to tryGetFundsFromRemoteCommitmentTx
Also pass a `increaseFeeToRaiseCpfpIncentive` argument to both
fund-recovery functions so that CPFP can be explicitly requested. If the
commitment tx is lingering in the mempool then this can be set to `true`
to increase the chance of it being mined. Otherwise, if the commitment
tx has already been mined then this can be set to `false` to avoid
paying a larger fee.