Upgrade to NBitcoin-5.0.75
This version enables setting transaction input sequence numbers and the
transaction version number through TransactionBuilder

Also Fix commitment number discovery in
tryGetFundsFromLocalCommitmentTx. This function would erroneously assume
that we are the funder and would fail to recover funds if the fundee
tried to use it.

Also amend the function to use the new sequence number and version
setting capabilities of NBitcoin.