Fix commitment number discovery bug.
Fix commitment number discovery in tryGetFundsFromLocalCommitmentTx.
This function would erroneously assume that we are the funder and would
fail to recover funds if the fundee tried to use it.

Also upgrade NBitcoin to v5.0.75 and amend
tryGetFundsFromLocalCommitmentTx to use the new sequence number and
version setting capabilities of the new version.

Co-authored-by: Afshin Arani <afshinarani79@gmail.com>