Move some channel fields into new StaticChannelConfig type
This type represents all channel configuration which is established
during the handshake and remains unchanged through the lifetime of the
channel.

The following fields are moved into this type: AnounceChannel,
RemoteNodeId, Network, FundingTxMinimumDepth, the shutdown scripts,
IsFunder, FundingScriptCoin, {Local,Remote}Params, and
RemoteChannelPubKeys. In doing so, we reduce the number of fields in
Commitments and the channel types.

Also, in CommitmentsModule, rather than passing around all the fields
individually pass the entire StaticChannelConfig object.